+++ scripts/main.js
@@ -12,6 +12,7 @@ import { registerItemSheetHooks } from "./item-sheet.js";
 import { registerItemChatHooks } from "./item-chat.js";
 import { registerProfessionHooks } from "./professions.js";
 import { registerWorkbench } from "./workbench.js";
+import { registerGMItemCreator } from "./gm-item-creator.js";
 
 let bridgeInstalled = false;
 let abilityBoostsInstalled = false;
@@ -52,6 +53,7 @@ Hooks.once("init", () => {
     restricted: false,
   });
   workbenchInstalled = true;
+  registerGMItemCreator();
 
   bridgeInstalled = installRuleElementBridge(getRulesConfig);
   abilityBoostsInstalled = installAbilityBoostBridge();